+++ tests/pallas/tpu_ops_test.py
@@ -1251,6 +1251,36 @@ class OpsTest(ptu.PallasTPUTest):
     expected = op(x, y)
     self.assertAllClose(out, expected)
 
+  @parameterized.product(approx=[False, True], full_range=[False, True], dtype=[jnp.bfloat16, jnp.float32])
+  def test_reciprocal(self, approx, full_range, dtype):
+    if dtype == jnp.bfloat16:
+      if not jtu.is_libtpu_at_least("0.0.46"):
+        self.skipTest("BF16 reciprocal requires libtpu >= 0.0.46.")
+      if not jtu.is_device_tpu_at_least(6):
+        self.skipTest("BF16 reciprocal not supported before v6e.")
+      if approx:
+        self.skipTest("BF16 reciprocal does not support approx=True.")
+      if not full_range:
+        self.skipTest("BF16 reciprocal does not support full_range=False.")
+
+    shape = (32, 256)
+    x = jnp.arange(np.prod(shape), dtype=dtype).reshape(shape)
+    if not full_range:
+      x = jnp.where(x == 0, -1.0, x)
+
+    def kernel(x_ref, o_ref):
+      o_ref[...] = pltpu.reciprocal(
+          x_ref[...], approx=approx, full_range=full_range
+      )
+
+    out = self.pallas_call(
+        kernel, out_shape=jax.ShapeDtypeStruct(shape, dtype)
+    )(x)
+    kwargs = {}
+    if approx:
+      kwargs.update(dict(atol=2e-5, rtol=2e-5))
+    np.testing.assert_allclose(out, jax.lax.reciprocal(x), **kwargs)
+
 
 class ConvTest(ptu.PallasTPUTest):